retrieval-agent/requirements.txt
# Core dependencies for Slice 1: Exa Integration
pydantic==2.5.0
exa-py==1.0.9
httpx==0.25.2
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

